move classes into Saul
