+++ src/main/java/org/fundacionjala/prog101/Saul/Research/Car.java
@@ -1,4 +1,4 @@
-package org.fundacionjala.prog101.Research;
+package org.fundacionjala.prog101.Saul.Research;
 
 public class Car {
     private String brand;
