+++ src/main/java/org/fundacionjala/prog101/Saul/Wrapper/IntWrapper.java
@@ -1,4 +1,4 @@
-package org.fundacionjala.prog101.Wrapper;
+package org.fundacionjala.prog101.Saul.Wrapper;
 
 public final class IntWrapper {
     private final int value;
